extract more code to different files & some code optimization
