+++ source/js/common/template-js-handles/footer-donate-button.js
@@ -2,6 +2,7 @@ import { ReactGA } from "../react-ga-proxy.js";
 
 /**
  * Bind click handler to #donate-footer-btn
+ * ("Donate" button on footer)
  */
 export default () => {
   let donateFooterBtn = document.getElementById(`donate-footer-btn`);
