+++ source/js/common/template-js-handles/header-donate-button.js
@@ -1,0 +1,18 @@
+import { ReactGA } from "../react-ga-proxy.js";
+
+/**
+ * Bind click handler to #donate-header-btn
+ * ("Donate" button on primary nav)
+ */
+export default () => {
+  let donateHeaderBtn = document.getElementById(`donate-header-btn`);
+  if (donateHeaderBtn) {
+    donateHeaderBtn.addEventListener(`click`, () => {
+      ReactGA.event({
+        category: `donate`,
+        action: `donate button tap`,
+        label: `${document.title} header`
+      });
+    });
+  }
+};
