+++ source/js/common/template-js-handles/index.js
@@ -1,8 +1,10 @@
-import bindFooterDonateButtonHandler from "./footer-donate-button";
+import bindHeaderDonateButtonHanlder from "./header-donate-button.js"
+import bindFooterDonateButtonHandler from "./footer-donate-button.js";
 
 /**
  * Bind event handlers
  */
 export const bindCommonEventHandlers = () => {
+  bindHeaderDonateButtonHanlder();
   bindFooterDonateButtonHandler();
 };
