+++ source/js/main.js
@@ -11,11 +11,15 @@ import {
   ReactGA
 } from "./common";
 
-import { bindEventHandlers, injectReactComponents } from "./foundation";
+import {
+  bindGlobalEventHandlers,
+  bindEventHandlers,
+  injectReactComponents
+} from "./foundation";
+
+import { bindEventHandlers as bindMozFestEventHandlers } from "./mozfest";
 
 import primaryNav from "./primary-nav.js";
-import bindMozFestGA from "./mozfest-ga.js";
-import bindMozFestEventHandlers from "./mozfest-event-handlers";
 import youTubeRegretsTunnel from "./youtube-regrets.js";
 
 // Initializing component a11y browser console logging
@@ -62,8 +66,17 @@ let main = {
       GoogleAnalytics.init();
 
       this.injectReactComponents();
-      this.bindGlobalHandlers();
-      this.bindGAEventTrackers();
+      this.bindHandlers();
+
+      // bind MozFest specific script if on MozFest pages
+      if (document.querySelector("body").classList.contains("mozfest")) {
+        bindMozFestEventHandlers();
+      }
+
+      // initialize YouTube Regret interactive tunnel if on YouTube Regrets page
+      if (document.querySelector("#view-youtube-regrets")) {
+        youTubeRegretsTunnel.init();
+      }
 
       // Record that we're done, when we're really done.
       Promise.all(apps).then(() => {
@@ -83,154 +96,18 @@ let main = {
     envReq.send();
   },
 
-  fetchHomeDataIfNeeded(callback) {
-    // Only fetch data if you're on the homepage
-    if (document.querySelector(`#view-home`)) {
-      let homepageReq = new XMLHttpRequest();
-
-      homepageReq.addEventListener(`load`, () => {
-        callback.call(this, JSON.parse(homepageReq.response));
-      });
-
-      homepageReq.open(`GET`, `${networkSiteURL}/api/homepage/`);
-      homepageReq.send();
-    } else {
-      callback.call(this, {});
-    }
-  },
-
-  bindGlobalHandlers() {
-    // Track window scroll position and add/remove class to change sticky header appearance
-
-    let lastKnownScrollPosition = 0;
-    let ticking = false;
-    let elBurgerWrapper = document.querySelector(`.wrapper-burger`);
-
-    let adjustNavbar = scrollPosition => {
-      if (scrollPosition > 0) {
-        elBurgerWrapper.classList.add(`scrolled`);
-      } else {
-        elBurgerWrapper.classList.remove(`scrolled`);
-      }
-    };
-
-    let elCtaAnchor = document.querySelector(`#cta-anchor`);
-    let elStickyButton = document.querySelector(
-      `.narrow-sticky-button-container`
-    );
-    let noopCtaButton = () => {};
-    let adjustCtaButton = noopCtaButton;
-
-    if (elCtaAnchor && elStickyButton) {
-      let getAnchorPosition = () => {
-        return (
-          elCtaAnchor.getBoundingClientRect().top +
-          window.scrollY -
-          window.innerHeight
-        );
-      };
-
-      let ctaAnchorPosition = getAnchorPosition();
-
-      window.addEventListener(`resize`, () => {
-        ctaAnchorPosition = getAnchorPosition();
-      });
-
-      let scrollCtaButton = scrollPosition => {
-        if (scrollPosition > ctaAnchorPosition) {
-          elStickyButton.classList.add(`hidden`);
-          adjustCtaButton = noopCtaButton;
-        }
-      };
-
-      let initCtaButton = scrollPosition => {
-        if (scrollPosition <= ctaAnchorPosition) {
-          elStickyButton.classList.remove(`hidden`);
-          adjustCtaButton = scrollCtaButton;
-        }
-      };
-
-      adjustCtaButton = initCtaButton;
-    }
-
-    let onScroll = () => {
-      lastKnownScrollPosition = window.scrollY;
-
-      if (!ticking) {
-        window.requestAnimationFrame(() => {
-          adjustNavbar(lastKnownScrollPosition);
-          adjustCtaButton(lastKnownScrollPosition);
-          ticking = false;
-        });
-      }
-
-      ticking = true;
-    };
-
-    window.addEventListener(`scroll`, onScroll);
-
-    // Toggle sticky share buttons on blog page
-
-    let blogPageStickyButtons = document.querySelector(
-      `#view-blog .blog-sticky-side .share-button-group`
-    );
-    let blogPageFullButtons = document.querySelector(
-      `#view-blog .blog-body .share-button-group`
-    );
-
-    if (blogPageStickyButtons && blogPageFullButtons) {
-      const isInViewport = element => {
-        let box = element.getBoundingClientRect();
-
-        return box.top <= window.innerHeight && box.top + box.height >= 0;
-      };
-
-      const toggleStickyButtons = () => {
-        if (isInViewport(blogPageFullButtons)) {
-          blogPageStickyButtons.classList.add(`faded`);
-        } else {
-          blogPageStickyButtons.classList.remove(`faded`);
-        }
-      };
-
-      window.addEventListener(`scroll`, toggleStickyButtons);
-      toggleStickyButtons();
-    }
-
-    // Call once to get scroll position on initial page load.
-    onScroll();
+  bindHandlers() {
+    bindCommonEventHandlers();
+    bindGlobalEventHandlers();
+    bindEventHandlers();
 
     initializePrimaryNav(networkSiteURL, csrfToken, primaryNav);
-    youTubeRegretsTunnel.init();
-
-    // Extra tracking
-
-    let donateHeaderBtn = document.getElementById(`donate-header-btn`);
-    if (donateHeaderBtn) {
-      donateHeaderBtn.addEventListener(`click`, () => {
-        ReactGA.event({
-          category: `donate`,
-          action: `donate button tap`,
-          label: `${document.title} header`
-        });
-      });
-    }
-
-    bindCommonEventHandlers();
-  },
-
-  bindGAEventTrackers() {
-    // MozFest specific scripts
-    bindMozFestGA();
-    bindMozFestEventHandlers();
   },
 
   // Embed various React components based on the existence of containers within the current page
   injectReactComponents() {
     injectCommonReactComponents(apps, networkSiteURL, csrfToken);
     injectReactComponents(apps, networkSiteURL, env);
-
-    bindEventHandlers();
   }
 };
